feat: SHA512SUMS support
This adds the regen-sha512sums script to the repo which recalculates
SHA512 checksums for all files in the repo and signs them with a
PKCS#11 certificate. The trusted CAs for verification are included in
the data/trusted-ca.pem file.

All commits moving forward MUST update the SHA512SUMS and SHA512SUMS.p7s
files to ensure the integrity of the repo.

Signed-off-by: Jaremy Hatler <root@jhatler.com>