docs: add README with overview and OpSec details
This is the initial addition of the README.md file to this repository.
It provides an overview of its purpose, usage, and operational security
details. A canary statement has also been included at the top of the
file for use in the event of a compromise.

A link to my Keybase account was also included for verification
purposes.

Signed-off-by: Jaremy Hatler <root@jhatler.com>